Rotation fonctionne pour les poteaux et les murs sans vérifier la position sur la table.

+++ Cadriciel/Sources/DLL/Visiteur/VisiteurAbstrait.h
@@ -50,8 +50,8 @@ public:
 	std::shared_ptr<const NoeudAbstrait> getReferenceNoeud();
 
 	inline void assignerPositionRelative(glm::dvec3 positionRelative);
-	inline void assignerAngleRotation(float angleRotation);
-	inline void assignerFacteurDimension(float facteurDimension);
+	inline void assignerAngleRotation(double angleRotation);
+	inline void assignerFacteurDimension(double facteurDimension);
 	inline std::shared_ptr<NoeudAbstrait> obtenirReferenceNoeud();
 
 	virtual void visiter(ArbreRendu* noeud);
@@ -60,12 +60,10 @@ public:
 	virtual void visiter(NoeudMur* noeud);
 	virtual void visiter(NoeudLigneNoire* noeud);
 
-
-
 protected:
 	glm::dvec3 positionRelative_;
-	float angleRotation_;
-	float facteurDimension_;
+	double angleRotation_;
+	double facteurDimension_;
 	modele::Modele3D* referenceModele_;
 	std::shared_ptr<NoeudAbstrait> referenceNoeud_;
 };
@@ -77,12 +75,12 @@ inline void VisiteurAbstrait::assignerPositionRelative(glm::dvec3 positionRelati
 	positionRelative_ = positionRelative;
 }
 
-inline void VisiteurAbstrait::assignerAngleRotation(float angleRotation)
+inline void VisiteurAbstrait::assignerAngleRotation(double angleRotation)
 {
 	angleRotation_ = angleRotation;
 }
 
-inline void VisiteurAbstrait::assignerFacteurDimension(float facteurDimension)
+inline void VisiteurAbstrait::assignerFacteurDimension(double facteurDimension)
 {
 	facteurDimension_ = facteurDimension;
 }
@@ -91,7 +89,6 @@ inline std::shared_ptr<NoeudAbstrait> VisiteurAbstrait::obtenirReferenceNoeud()
 {
 	return referenceNoeud_;
 }
-
 
 #endif // VISITEUR_ABSTRAIT_H
